Add part of the name selection screen, get rid of some warnings

+++ utils.hpp
@@ -77,7 +77,7 @@ template <typename T, size_t N> void iota(Array<T, N>& array)
 template <typename T, size_t N> void shuffle(Array<T, N>& array)
 {
     for (size_t i = N - 1; i >= 1; --i)
-        Tiny::swap(array[i], array[random(i + 1)]);
+        Tiny::swap(array[i], array[size_t(random(i + 1))]);
 }
 
 template <typename T, typename U, size_t N> size_t find(const Array<T, N>& arr, const U& value)
